RAG_simples_2/pyproject.toml
[project]
name = "cap14"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv-python>=0.0.1",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-session==0.5.0",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
]
